Split variant_storage.hpp and visit.hpp
